requirements.txt
kfp
pandas
sklearn
google-cloud-aiplatform